extract more testing helper fns
